+++ back/src/types/express-session/index.d.ts
@@ -1,0 +1,14 @@
+import "express-session";
+
+declare module "express-session" {
+  export interface SessionData {
+    usuario?: {
+      id: number;
+      nome: string;
+      tipo: "aluno" | "funcionario";
+      matricula?: string;
+      cargo?: string;
+      curso?: string;
+    };
+  }
+}
