+++ front/horas-extracurriculares/package.json
@@ -11,6 +11,7 @@
     "postcss": "^8.5.6",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^6.30.1",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4",
     "webpack-dev-server": "^5.2.2"
